Move brewpi_spawner to utils directory to prevent runtime incompatibility with Celery
