+++ circus.ini
@@ -13,11 +13,11 @@ numprocesses = 2
 copy_env = True
 virtualenv = $(circus.env.HOME)/venv
 working_dir = $(circus.env.HOME)/fermentrack
-# Log ouput, rotate every 10MB, keep 5 logfiles.
+# Log ouput, rotate every 2MB, keep 5 logfiles.
 stderr_stream.class = FileStream
 stderr_stream.filename = $(circus.env.HOME)/fermentrack/log/fermentrack.log
 stderr_stream.time_format = %Y-%m-%d %H:%M:%S
-stderr_stream.max_bytes = 10485760
+stderr_stream.max_bytes = 2097152
 stderr_stream.backup_count = 5
 
 [socket:Fermentrack]
@@ -30,20 +30,22 @@ port = 8123
 #PYTHONPATH = $(circus.env.HOME)/fermentrack
 PYTHONPATH = /home/fermentrack/fermentrack
 
+
+
 [watcher:brewpi-spawner]
 cmd = $(circus.env.HOME)/venv/bin/python
-args = $(circus.env.HOME)/fermentrack/fermentrack_django/brewpi_spawner.py
+args = $(circus.env.HOME)/fermentrack/utils/brewpi_spawner.py
 virtualenv = $(circus.env.HOME)/venv
 working_dir = $(circus.env.HOME)/fermentrack
 copy_env = True
 
 use_sockets = False
 numprocesses = 1
-# Log ouput, rotate every 10MB, keep 5 logfiles.
+# Log ouput, rotate every 2MB, keep 5 logfiles.
 stderr_stream.class = FileStream
 stderr_stream.filename = $(circus.env.HOME)/fermentrack/log/fermentrack-brewpi-spawner.log
 stderr_stream.time_format = %Y-%m-%d %H:%M:%S
-stderr_stream.max_bytes = 10485760
+stderr_stream.max_bytes = 2097152
 stderr_stream.backup_count = 5
 
 
@@ -58,11 +60,12 @@ copy_path = True
 virtualenv = $(circus.env.HOME)/venv
 working_dir = $(circus.env.HOME)/fermentrack
 
+# Log ouput, rotate every 2MB, keep 5 logfiles.
 stderr_stream.class = FileStream
 stdout_stream.class = FileStream
 stdout_stream.filename = $(circus.env.HOME)/fermentrack/log/celery-stdout.log
 stderr_stream.filename = $(circus.env.HOME)/fermentrack/log/celery-stderr.log
 stdout_stream.refresh_time = 0.3
-stdout_stream.max_bytes = 1000000
+stdout_stream.max_bytes = 2097152
 stdout_stream.backup_count = 5
